04-fitness-coach-bot/requirements.txt
requests
rich
click
PyYAML
streamlit
pytest
